bundle_pipeline/requirements.txt
boto3
soundfile
openai-whisper
openai
pyyaml
qrcode[pil]
